Recommendation/requirement.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
docopt==0.6.2
Flask==2.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
python-decouple==3.6
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.2
yarg==0.1.9
